env variables fixed -> player registered successfully

+++ bsz-backend/src/main/java/org/dn/team/basszhattyuk/repository/dev/DevFileRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 @Repository
 @Profile("dev")
-public interface DevFileRepository extends JpaRepository<FileData, Integer> {
+public interface DevFileRepository extends JpaRepository<FileData, Long> {
 
     Optional<FileData> findByFileName(String fileName);
 }
